ml/ml_001/requirements.txt
scikit-learn
pandas
numpy
mlflow
pyyaml
matplotlib
seaborn
jupyter